Move CancellationTokenMustBeLast Analyzer to ApiDesignGuidelines analyer
