+++ src/Microsoft.ApiDesignGuidelines.Analyzers/Core/CancellationTokenParametersMustComeLast.cs
@@ -1,39 +1,116 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using System;
 using System.Collections.Immutable;
+using System.Linq;
 using Analyzer.Utilities;
-using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Diagnostics;
 
 namespace Microsoft.ApiDesignGuidelines.Analyzers
-{                   
+{
     /// <summary>
-    /// RS0011: CancellationToken parameters must come last
+    /// CA1068: CancellationToken parameters must come last.
     /// </summary>
-    public abstract class CancellationTokenParametersMustComeLastAnalyzer : DiagnosticAnalyzer
+    [DiagnosticAnalyzer(LanguageNames.CSharp, LanguageNames.VisualBasic)]
+    public sealed class CancellationTokenParametersMustComeLastAnalyzer : DiagnosticAnalyzer
     {
-        internal const string RuleId = "RS0011";
+        internal const string RuleId = "CA1068";
 
         private static readonly LocalizableString s_localizableTitle = new LocalizableResourceString(nameof(MicrosoftApiDesignGuidelinesAnalyzersResources.CancellationTokenParametersMustComeLastTitle), MicrosoftApiDesignGuidelinesAnalyzersResources.ResourceManager, typeof(MicrosoftApiDesignGuidelinesAnalyzersResources));
-        
         private static readonly LocalizableString s_localizableMessage = new LocalizableResourceString(nameof(MicrosoftApiDesignGuidelinesAnalyzersResources.CancellationTokenParametersMustComeLastMessage), MicrosoftApiDesignGuidelinesAnalyzersResources.ResourceManager, typeof(MicrosoftApiDesignGuidelinesAnalyzersResources));
-        private static readonly LocalizableString s_localizableDescription = new LocalizableResourceString(nameof(MicrosoftApiDesignGuidelinesAnalyzersResources.CancellationTokenParametersMustComeLastDescription), MicrosoftApiDesignGuidelinesAnalyzersResources.ResourceManager, typeof(MicrosoftApiDesignGuidelinesAnalyzersResources));
-        
-        internal static DiagnosticDescriptor Rule = new DiagnosticDescriptor(RuleId,
-                                                                             s_localizableTitle,
-                                                                             s_localizableMessage,
-                                                                             DiagnosticCategory.ApiDesign,
-                                                                             DiagnosticSeverity.Warning,
-                                                                             isEnabledByDefault: false,
-                                                                             description: s_localizableDescription,
-                                                                             helpLinkUri: null,     // TODO: add MSDN url
-                                                                             customTags: WellKnownDiagnosticTags.Telemetry);
 
-        public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create(Rule);
+        internal static DiagnosticDescriptor Rule = new DiagnosticDescriptor(
+            RuleId,
+            s_localizableTitle,
+            s_localizableMessage,
+            DiagnosticCategory.Design,
+            DiagnosticSeverity.Warning,
+            isEnabledByDefault: true);
 
-        public override void Initialize(AnalysisContext analysisContext)
-        { 
-            
+        public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics { get { return ImmutableArray.Create(Rule); } }
+
+        public override void Initialize(AnalysisContext context)
+        {
+            context.RegisterCompilationStartAction(compilationContext =>
+            {
+                var cancellationTokenType = compilationContext.Compilation.GetTypeByMetadataName("System.Threading.CancellationToken");
+                if (cancellationTokenType != null)
+                {
+                    compilationContext.RegisterSymbolAction(symbolContext =>
+                    {
+                        var methodSymbol = (IMethodSymbol)symbolContext.Symbol;
+                        if (methodSymbol.IsOverride
+                            || methodSymbol.ExplicitInterfaceImplementations.Any()
+                            || ImplementsAnInterfaceMethodImplicitly(methodSymbol))
+                        {
+                            return;
+                        }
+
+                        var last = methodSymbol.Parameters.Length - 1;
+                        if (last >= 0 && methodSymbol.Parameters[last].IsParams)
+                        {
+                            last--;
+                        }
+
+                        // Skip optional parameters, UNLESS one of them is a CancellationToken
+                        // AND it's not the last one.
+                        if (last >= 0 && methodSymbol.Parameters[last].IsOptional
+                            && !methodSymbol.Parameters[last].Type.Equals(cancellationTokenType))
+                        {
+                            last--;
+
+                            while (last >= 0 && methodSymbol.Parameters[last].IsOptional)
+                            {
+                                if (methodSymbol.Parameters[last].Type.Equals(cancellationTokenType))
+                                {
+                                    symbolContext.ReportDiagnostic(Diagnostic.Create(
+                                        Rule, methodSymbol.Locations.First(), methodSymbol.ToDisplayString()));
+                                }
+
+                                last--;
+                            }
+                        }
+
+                        while (last >= 0 && methodSymbol.Parameters[last].RefKind != RefKind.None)
+                        {
+                            last--;
+                        }
+
+                        for (int i = last; i >= 0; i--)
+                        {
+                            var parameterType = methodSymbol.Parameters[i].Type;
+                            if (parameterType.Equals(cancellationTokenType)
+                                && i != last)
+                            {
+                                symbolContext.ReportDiagnostic(Diagnostic.Create(
+                                    Rule, methodSymbol.Locations.First(), methodSymbol.ToDisplayString()));
+                                break;
+                            }
+                        }
+                    },
+                    SymbolKind.Method);
+                }
+            });
+        }
+
+        private bool ImplementsAnInterfaceMethodImplicitly(IMethodSymbol methodSymbol)
+        {
+            // This is an approximation, because another class could derive from this one
+            // and rely on methodSymbol implementing one of *it's* interfaces methods, but
+            // it's good enough.
+            foreach (var interfaceSymbol in methodSymbol.ContainingType.AllInterfaces)
+            {
+                foreach (var interfaceMethod in interfaceSymbol.GetMembers().Where(m => m.Kind == SymbolKind.Method))
+                {
+                    if (methodSymbol.ContainingType.FindImplementationForInterfaceMember(interfaceMethod)?.Equals(methodSymbol) ?? false)
+                    {
+                        return true;
+                    }
+                }
+            }
+
+            return false;
         }
     }
 }
