+++ src/Microsoft.ApiDesignGuidelines.Analyzers/Core/Microsoft.ApiDesignGuidelines.Analyzers.csproj
@@ -161,7 +161,6 @@
     <Compile Include="ImplementIEquatableWhenOverridingObjectEquals.cs" />
     <Compile Include="ImplementIEquatableWhenOverridingObjectEquals.Fixer.cs" />
     <Compile Include="CancellationTokenParametersMustComeLast.cs" />
-    <Compile Include="CancellationTokenParametersMustComeLast.Fixer.cs" />
     <Compile Include="DoNotDeclareStaticMembersOnGenericTypes.cs" />
     <Compile Include="DoNotDeclareStaticMembersOnGenericTypes.Fixer.cs" />
     <Compile Include="CollectionsShouldImplementGenericInterface.cs" />
