+++ src/Microsoft.ApiDesignGuidelines.Analyzers/UnitTests/Microsoft.ApiDesignGuidelines.Analyzers.UnitTests.csproj
@@ -207,7 +207,6 @@
     <Compile Include="ImplementIEquatableWhenOverridingObjectEqualsTests.cs" />
     <Compile Include="ImplementIEquatableWhenOverridingObjectEqualsTests.Fixer.cs" />
     <Compile Include="CancellationTokenParametersMustComeLastTests.cs" />
-    <Compile Include="CancellationTokenParametersMustComeLastTests.Fixer.cs" />
     <Compile Include="DoNotDeclareStaticMembersOnGenericTypesTests.cs" />
     <Compile Include="DoNotDeclareStaticMembersOnGenericTypesTests.Fixer.cs" />
     <Compile Include="CollectionsShouldImplementGenericInterfaceTests.cs" />
