+++ src/Microsoft.ApiDesignGuidelines.Analyzers/VisualBasic/Microsoft.ApiDesignGuidelines.VisualBasic.Analyzers.vbproj
@@ -136,8 +136,6 @@
     <Compile Include="BasicPassSystemUriObjectsInsteadOfStrings.Fixer.vb" />
     <Compile Include="BasicImplementIEquatableWhenOverridingObjectEquals.vb" />
     <Compile Include="BasicImplementIEquatableWhenOverridingObjectEquals.Fixer.vb" />
-    <Compile Include="BasicCancellationTokenParametersMustComeLast.vb" />
-    <Compile Include="BasicCancellationTokenParametersMustComeLast.Fixer.vb" />
     <Compile Include="BasicDoNotDeclareStaticMembersOnGenericTypes.vb" />
     <Compile Include="BasicDoNotDeclareStaticMembersOnGenericTypes.Fixer.vb" />
     <Compile Include="BasicCollectionsShouldImplementGenericInterface.Fixer.vb" />
